Fix audio judder etc
A really silly mistake from me: I forgot that the ValueChanged event on sliders fires when the upstream data-binding updates.  In other words, it also fires when the server mentions a marker position change---causing the UI to send the server a request to change to the position it just broadcast!  Due to the round-trip latency, this causes the server to skip back a little, hence judder.