Command refactor: stable state
A lot of the command decoder now uses unpacked commands, and is generally a lot tidier than before.

There's still a lot to do here to clean up command decoding, as well as to generalise the command decoder for server usage too.  But it passes what few tests it has, so I'm calling it for now.