Tests: factor out event setup, fix? heisenbug
I'm going to start rolling out unit tests for the other view models (hence the empty DirectoryViewModelTests class), and so have factored out the somewhat unpleasant event feed creation into an abstract base class.

While doing so, I realised there's a fairly awful heisenbug in the text tests, which seems to be caused by trying to check the CanExecute of a command before it's finished executing.  (No, I don't know why this is happening in a seemingly sequential test case).  I've tried to fix by forcing the check to first make sure the command has stopped executing, but I'm not definitively sure as to whether this has worked yet.